backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
praw>=7.7.0
supabase>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
textblob>=0.17.1
vaderSentiment>=3.3.2
authlib>=1.2.1
itsdangerous>=2.1.2
starlette>=0.27.0
